Cities admin

+++ app/controllers/cities_controller.rb
@@ -21,6 +21,7 @@ class CitiesController < ApplicationController
     return if inactive_redirect
     @city = City.find_by_id(params[:id])
     posts = Post.all
+    posts = posts.select{ |post| post[:city_id] == @city[:id]}
     @active_user_posts = []
     posts.each do |post|
       if post.user[:active]
@@ -77,10 +78,7 @@ class CitiesController < ApplicationController
 
 
   def destroy
-    if @current_user[:username] != "admin"
-      redirect_to cities_path
-      flash[:error] = "You're not authorized to delete a city."
-    else
+    if current_user && current_user.username == "admin"
       @city = City.find_by_id(params[:id])
       if @city.destroy
         flash[:success] = "The city has been deleted."
@@ -89,6 +87,9 @@ class CitiesController < ApplicationController
         flash[:error] = @city.errors.full_messages.join(", ")
         redirect_to edit_city_path(@city)
       end
+    else 
+      redirect_to cities_path
+      flash[:error] = "You're not authorized to delete a city."
     end
   end
